requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
faiss-cpu==1.12.0
filetype==1.2.0
frozenlist==1.8.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-ollama==0.3.8
langchain-openai==0.3.32
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.26
lxml==6.0.2
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.4
olefile==0.47
ollama==0.6.0
openai==1.106.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
propcache==0.4.1
psutil==7.1.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pypdf==6.1.3
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sseclient==0.0.27
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.56.2
trl==0.22.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unsloth==2025.10.10
unsloth_zoo==2025.10.12
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
wrapt==2.0.0
xxhash==3.5.0
yarl==1.22.0
zstandard==0.24.0
